Make the Context Hierarchy a full Directed Acyclic Graph
Until now the context hierarchy, that is, the relationship between
contexts, was a simple tree. Every context had just a single
parent. While this hierarchy model applies to a lot of use cases and
applications, it does not cover all of them.

Recently we have had the need to use the RBAC library in a couple of
projects where a simple tree was not enough, and we needed a full
Directed Acyclic Graph (DAG) to model the context hierarchy. The
reason for that was that a given context needed to be related to more
than one parent context at the same time.

This also means that we can also have multiple independent root
contexts (contexts that don't have *any* ancestors).

As a consequence, the underlying database model has changed. A new
table is used (`rbac_context_parent`), and the parent-child
relationship information needs to be moved from the `rbac_context`
table to the new `rbac_context_parent` table.