Remove all pre-condition assertions
It turns out we don't get as good diagnostics out of them as we'd
like. We'd like to know not just what predicate fails, but why. spec
can give us that, and we already have the specs for all the public
functions.

Also, asserting in a library is a controversial issue in the Clojure
community. See, e.g.,
https://gist.github.com/daveliepmann/8289f0ee5b00a5f05b50379e07fceb76

Thus we have decided that it is up to the consumer of the library to
decide if s/he wants the detailed information that spec can provide or
not (because checking the spec and getting the detailed information
has a performance hit).

Consumers interested in checking the args with spec, and getting
detailed diagnostics in case the args are not up to what is expected
of them, can use `clojure.spec.test.alpha/instrument` to instrument
the relevant functions they are interested in.

Or they can use `dev.gethop.rbac.next/instrument` to instrument all
the functions in the library namespace (that have a spec, of course),
except for `dev.gethop.rbac.next/has-permission?` (for performance
reasons). There is also the complementary function
`dev.gethop.rbac.next/unstrument`.