Use version 7 UUIDs by default, but make it configurable.
Commit ebb409bf8f started using version 7 UUIDs, instead of version 4
UUIDs, for better database performance (indices behaviour is much
better). In this case, leaking time information from the UUIDs should
not be an issue.

The new `set-primary-key-fn` function allows the user of the library
to override how the primary keys values are created. Thus letting the
caller use completely different database types, etc. (Issue #4)

The new `set-primary-key-fn` function allows the user of the library
to override how the primary keys managed by the library itself are
created. That the library lets the caller use completely different
database types, or even workarounds to use UUIDs for databases that do
not support them natively (like sqlite, or MySQL).

[Re: #4]