Support deleting a context using its id
While the function spec for `delete-context!` clearly says that you
have to provide a full context definition as returned by
`create-context!`, the actual implementation only used the
:resource-id and :conext-type-name keys.

Now some users want to be able to delete a context by just passing
its id, instead of the full context definition. But because the
current implementation doesn't use that key, the function doesn't
delete the context.

We have decided that in order to keep backwards compatibility we are
going to change the implementation to use the :id key if it's
present (it should be, if the caller respected the function spec), and
fall back to the :resource-id and :context-type-name keys if it's not.

In addition to that, we are going to introduce two new functions
called `delete-context-by-id` and `delete-contexts-by-ids`, just like
we have `delete-permission-by-id`, `delete-role-by-id`, etc, that are
100% explicit about taking id, and only ids.

[Re: #1]