Use version 7 UUIDs
[Re #4]

We decided to switch from UUID v4 (default generated by JDK) to UUID
v7 using the clj-uuid library. Mainly for its advantages in a
relational database model in PostgreSQL. v4s are completely random
which is not good for indexes as there is no order and thus indexes
that should be close to each other can be in completely separate
locations. This locality issue causes performance issues because
PostgreSQL might need several pages in the index cache and do way more
page checks because the indexes are disperse. Apart from that, the
statistics on random UUIDs is very bad because there is no lexical
ordering, so the planner can't tell if using an index is bad or good
at a giving moment. That causes the planner to make strange decisions
at times and the only way to improve it slightly is to increase the
amount of samples ANALYZE can use for a given table column that uses
UUID as data type. With more samples (buckets) it helps PostgreSQL to
have a better idea of the data distribution but still it's completely
random and the indexes issue continues. v7 however has lexical
ordering as part of it is based on unix timestamp which is already way
better because now we can be sure that records inserted together will
most likely be together in the same index page. It will also reduce
the amount pages in the cache and at the same time keeping the extra
randomness we need from UUIDs.

Even if the library is not PostgreSQL specific, we think that it will
benefit other database engines too.